requirements.txt
Django==3.2.10
Markdown===3.3.3